Move loan commands to command package
